observability: make metrics-port assignment a single source of truth
Standalone `cluster prometheus-config` rendered per-component default ports
(master 9324 / volume 9325 / filer 9326) when metrics_port was unset, while
the integrated deploy assigned per-host-unique ports from 9324 and fed those
to weed's -metricsPort. The two could disagree, so a hand-run scrape config
wouldn't match what weed actually served.

Move assignMetricsPorts into the spec package as spec.AssignMetricsPorts (the
single source of truth — idempotent, preserves explicit ports) and run it from
both DeployCluster and RenderPromConfig. Drop the divergent
Default{Master,Volume,Filer}MetricsPort constants. The rendered scrape config
now always matches the ports weed is started with.

Tests: move the assignment tests to the spec package, add idempotency and a
co-located-distinct-ports render test, and update the golden config (a lone
volume now scrapes on 9324, not 9325).
