+++ pkg/cluster/manager/deploy_monitoring.go
@@ -12,79 +12,6 @@ import (
 	"golang.org/x/sync/errgroup"
 )
 
-// metricsPortBase is the first port assigned to a component's Prometheus
-// metrics endpoint when one isn't set explicitly.
-const metricsPortBase = 9324
-
-// assignMetricsPorts gives every master/volume/filer without an explicit
-// metrics_port a per-host-unique one (so two volume servers on the same
-// host don't collide), starting at metricsPortBase. Explicitly-set ports
-// are preserved and reserved. The same spec drives both the deploy
-// (-metricsPort) and the rendered Prometheus scrape config, so they stay
-// in sync. Only called when monitoring is enabled.
-func assignMetricsPorts(s *spec.Specification) {
-	if s == nil {
-		return
-	}
-	used := map[string]map[int]bool{}
-	mark := func(ip string, p int) {
-		if p == 0 {
-			return
-		}
-		if used[ip] == nil {
-			used[ip] = map[int]bool{}
-		}
-		used[ip][p] = true
-	}
-	for _, m := range s.MasterServers {
-		if m != nil {
-			mark(m.Ip, m.MetricsPort)
-		}
-	}
-	for _, v := range s.VolumeServers {
-		if v != nil {
-			mark(v.Ip, v.MetricsPort)
-		}
-	}
-	for _, f := range s.FilerServers {
-		if f != nil {
-			mark(f.Ip, f.MetricsPort)
-		}
-	}
-
-	next := map[string]int{}
-	alloc := func(ip string) int {
-		if used[ip] == nil {
-			used[ip] = map[int]bool{}
-		}
-		p := next[ip]
-		if p < metricsPortBase {
-			p = metricsPortBase
-		}
-		for used[ip][p] {
-			p++
-		}
-		used[ip][p] = true
-		next[ip] = p + 1
-		return p
-	}
-	for _, m := range s.MasterServers {
-		if m != nil && m.MetricsPort == 0 {
-			m.MetricsPort = alloc(m.Ip)
-		}
-	}
-	for _, v := range s.VolumeServers {
-		if v != nil && v.MetricsPort == 0 {
-			v.MetricsPort = alloc(v.Ip)
-		}
-	}
-	for _, f := range s.FilerServers {
-		if f != nil && f.MetricsPort == 0 {
-			f.MetricsPort = alloc(f.Ip)
-		}
-	}
-}
-
 type monHost struct {
 	ip   string
 	port int
